requirements.txt
torch==1.5.0+cu101
torchvision==0.6.0+cu101
Pillow==7.1.2